Nesting event listeners (callback hell)
A way to control the order in which the data from the AJAX calls comes in is explored. The event listeners that provide the asynchronicity for the AJAX calls can be nested, meaning one has to finishing loading before the nested one can even start listening for it's load event. But this solution can lead to what is known as callback hell.