Promises with fetch
The functionality of the app is essentially identical to the way it was before. But the architecture of the code was changed. Instead of using XML-HTTP-Request with events and callbacks to achieve asynchronicity. The fetch API/method was used with promises to achieve the same functionality minus the callback hell.