Challenge solution
A solution to the coding challenge that Jonas created for the asynchronous JavaScript section. The solution i provided with no know issues, however, the second part of it is different from Jonas's solution. I simply plugged the function responsible for rendering the country cards on screen into the "whereAmI" function from the challenge, and the country was passed into the rendering function with data from the goecoding API. 

Jonas's solution involved re-writing or copying all the code contained in the rendering function, and then slightly modifying it to work with the "whenAmI" function. I believe my solution is better as in adheres to the DRY principle, and i more inline with the lazy nature of the programmer.