Short note + small changes
A short note on the JavaScript Runtime Environment, and it's event loop in particular. A small change was also made to the way the population number is displayed in the UI. It now rounds off to there nearest million, and has two toFixed() decimals after. When not so tired i shall think of a way to do this programmatically, as there are countries will more less than a million people like Barbados, and countries with more than 999 million, like India.