Small Bug Fix
There was some buggy behaviour coming from the "showCountry" async function. The bug was due to the entire function (including it's declaration) being wrapped in the try block, when it should have been only the function body to be wrapped in the try block! It's now been fixed, and the function can now be called from outside the try block!