Running Promises in Parallel
In order to make multiple calls to the rest-countries api endpoint simultaneously, since the data in this particular implementation doesn't depend on each other the Promise.all combinator function was used. This function creates a Promise that is resolved with an array of results when all of the provided Promises resolve, or rejected when any of the Promises are rejected.  It's real nice, and they are other!